chore: clean comments & kernels (#221)
Closes INT-4737 
Closes INT-4736
Relates to INT-4742

Should be done together with INT-4742 (delete unused kernels) to keep
track of actual changes in vendor files

+++ src/fri_log_up.rs
@@ -341,7 +341,7 @@ impl FriLogUpPhaseGpu {
                 rules,
                 used_nodes,
                 partition_lens,
-                partition_lens.len() as u32,
+                partition_lens.len(),
                 permutation_height as u32,
                 permutation_width_ext as u32,
                 tile_per_thread,
